Migrate more label endpoints, remove unused service method
