--- server/service/endpoint_labels.go
@@ -1,51 +1,0 @@
-package service
-
-import (
-	"context"
-
-	"github.com/fleetdm/fleet/v4/server/fleet"
-	"github.com/go-kit/kit/endpoint"
-)
-
-////////////////////////////////////////////////////////////////////////////////
-// Get Label Spec
-////////////////////////////////////////////////////////////////////////////////
-
-type getLabelSpecsResponse struct {
-	Specs []*fleet.LabelSpec `json:"specs"`
-	Err   error              `json:"error,omitempty"`
-}
-
-func (r getLabelSpecsResponse) error() error { return r.Err }
-
-func makeGetLabelSpecsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		specs, err := svc.GetLabelSpecs(ctx)
-		if err != nil {
-			return getLabelSpecsResponse{Err: err}, nil
-		}
-		return getLabelSpecsResponse{Specs: specs}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Get Label Spec
-////////////////////////////////////////////////////////////////////////////////
-
-type getLabelSpecResponse struct {
-	Spec *fleet.LabelSpec `json:"specs,omitempty"`
-	Err  error            `json:"error,omitempty"`
-}
-
-func (r getLabelSpecResponse) error() error { return r.Err }
-
-func makeGetLabelSpecEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(getGenericSpecRequest)
-		spec, err := svc.GetLabelSpec(ctx, req.Name)
-		if err != nil {
-			return getLabelSpecResponse{Err: err}, nil
-		}
-		return getLabelSpecResponse{Spec: spec}, nil
-	}
-}
