--- server/service/service_labels.go
@@ -1,31 +1,0 @@
-package service
-
-import (
-	"context"
-
-	"github.com/fleetdm/fleet/v4/server/fleet"
-)
-
-func (svc *Service) GetLabelSpecs(ctx context.Context) ([]*fleet.LabelSpec, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.GetLabelSpecs(ctx)
-}
-
-func (svc *Service) GetLabelSpec(ctx context.Context, name string) (*fleet.LabelSpec, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.GetLabelSpec(ctx, name)
-}
-
-func (svc *Service) ListLabelsForHost(ctx context.Context, hid uint) ([]*fleet.Label, error) {
-	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionRead); err != nil {
-		return nil, err
-	}
-
-	return svc.ds.ListLabelsForHost(ctx, hid)
-}
